Display quotas per Cinder volume type
This patch makes Cinder volume type quotas visible in panels
related to volume creation (from volumes and images tabs).

Currently user is able to select a desired type, but if quotas are
set, there is no way to check if creating a volume is going to
succeed. With this patch two additional quota charts are displayed
so that end user can see if there is still available quota for
selected type.

Closes-Bug: #1717342
Implements: blueprint quotas-per-cinder-volume-type

Change-Id: I5099cc442f023c7a67203ccd3cf647cca9bdb52c